gpu_base/requirements.txt
torch
transformers
pandas
pyyaml
tqdm
